:construction: fixed ft_error
*

*

+++ inc/cub3d_game.h
@@ -6,7 +6,7 @@
 /*   By: sshakya <sshakya@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/06 23:04:14 by sshakya           #+#    #+#             */
-/*   Updated: 2021/04/13 18:45:21 by sshakya          ###   ########.fr       */
+/*   Updated: 2021/04/23 04:43:26 by sshakya          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ typedef struct		s_player
 {
 	t_vector		vector;
 	t_camera		camera;
-	uint_fast8_t	move;
+	uint_fast16_t	move;
 }					t_player;
 /*
 ** WORLD -- SPRITE, PARAMS, PATHS
